Final clean up of EDM development.
